requirements.txt
# NexaVault – Python dependencies
# Install with: pip install -r requirements.txt

# ── Algorand ─────────────────────────────────────────────────────────────────
algosdk>=2.4.0
pyteal>=0.20.1

# ── DEX ──────────────────────────────────────────────────────────────────────
tinyman-py-sdk>=2.0.0

# ── Web API ──────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── HTTP client (async) ───────────────────────────────────────────────────────
httpx>=0.27.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0       # load .env files

# ── Optional: ML strategy ─────────────────────────────────────────────────────
# Uncomment when plugging in FinRL or a custom sklearn model:
# finrl>=0.3.6
# scikit-learn>=1.4.0
# numpy>=1.26.0
# pandas>=2.0.0
